controller-grpc: Move pkg into controller/grpc and run as a process in the controller app
