+++ builder/manifest.json
@@ -147,6 +147,7 @@
       "layers": [{
         "gobuild": {
           "controller":           "/bin/flynn-controller",
+          "controller/grpc":      "/bin/flynn-controller-grpc",
           "controller/scheduler": "/bin/flynn-scheduler",
           "controller/worker":    "/bin/flynn-worker"
         },
@@ -162,35 +163,12 @@
         "run": [
           "mkdir -p /etc/flynn-controller",
           "cp -r schema /etc/flynn-controller/jsonschema"
-        ]
-      }],
-      "entrypoint": {
-        "args": [
-          "/bin/start-flynn-controller"
-        ]
-      }
-    },
-    {
-      "id": "controller-grpc",
-      "base": "busybox",
-      "layers": [{
-        "gobuild": {
-          "controller-grpc": "/bin/flynn-controller-grpc"
-        },
-        "inputs": [
-          "schema/*.json",
-          "schema/controller/*.json",
-          "schema/router/*.json"
-        ],
-        "run": [
-          "mkdir -p /etc/flynn-controller",
-          "cp -r schema /etc/flynn-controller/jsonschema"
         ],
         "limits": { "temp_disk": "1G" }
       }],
       "entrypoint": {
         "args": [
-          "/bin/flynn-controller-grpc"
+          "/bin/start-flynn-controller"
         ]
       }
     },
